Formatting/space/print statement
Changed "PA" to 'PA' in if statement and pause function to fit with the formatting of the rest of the code

Added another space to info in the pause function (Not sure if that will functionally change anything) since the following line checks for (I assume) the first two characters of the string, when info starts with only one, in that instance. 

Added print statement to visually show that the while loop in the pause function terminates.